+++ src/main/java/edu/ie3/datamodel/io/processor/Processor.java
@@ -15,6 +15,7 @@ import edu.ie3.datamodel.models.input.OperatorInput;
 import edu.ie3.datamodel.models.input.connector.SwitchInput;
 import edu.ie3.datamodel.models.input.system.characteristic.CharacteristicInput;
 import edu.ie3.datamodel.models.profile.LoadProfile;
+import edu.ie3.datamodel.models.result.CongestionResult;
 import edu.ie3.datamodel.models.voltagelevels.VoltageLevel;
 import edu.ie3.datamodel.utils.Try;
 import edu.ie3.datamodel.utils.Try.*;
@@ -293,6 +294,8 @@ public abstract class Processor<T> {
           "ReactivePowerCharacteristic",
           "CharacteristicInput" -> resultStringBuilder.append(
           ((CharacteristicInput<?, ?>) methodReturnObject).serialize());
+      case "InputModelType" -> resultStringBuilder.append(
+          ((CongestionResult.InputModelType) methodReturnObject).type);
       default -> throw new EntityProcessorException(
           "Unable to process value for attribute/field '"
               + fieldName